Remove controller SwiftUI additions
