--- Sources/StreamChat/Controllers/ChannelListController/ChannelListController+SwiftUI.swift
@@ -1,47 +1,0 @@
-//
-// Copyright © 2025 Stream.io Inc. All rights reserved.
-//
-
-import Foundation
-import SwiftUI
-
-extension ChatChannelListController {
-    /// A wrapper object that exposes the controller variables in the form of `ObservableObject` to be used in SwiftUI.
-    public var observableObject: ObservableObject { .init(controller: self) }
-
-    /// A wrapper object for `ChannelListController` type which makes it possible to use the controller comfortably in SwiftUI.
-    public class ObservableObject: SwiftUI.ObservableObject {
-        /// The underlying controller. You can still access it and call methods on it.
-        public let controller: ChatChannelListController
-
-        /// The channels matching the query.
-        @Published public private(set) var channels: [ChatChannel] = []
-
-        /// The current state of the Controller.
-        @Published public private(set) var state: DataController.State
-
-        /// Creates a new `ObservableObject` wrapper with the provided controller instance.
-        init(controller: ChatChannelListController) {
-            self.controller = controller
-            state = controller.state
-
-            controller.multicastDelegate.add(additionalDelegate: self)
-            channels = controller.channels
-        }
-    }
-}
-
-extension ChatChannelListController.ObservableObject: ChatChannelListControllerDelegate {
-    public func controller(
-        _ controller: ChatChannelListController,
-        didChangeChannels changes: [ListChange<ChatChannel>]
-    ) {
-        // We don't care about detailed changes. We just need to update the `channels` property and keep SwiftUI
-        // deal with the rest.
-        channels = controller.channels
-    }
-
-    public func controller(_ controller: DataController, didChangeState state: DataController.State) {
-        self.state = state
-    }
-}
