--- Sources/StreamChat/Controllers/ChannelWatcherListController/ChatChannelWatcherListController+SwiftUI.swift
@@ -1,47 +1,0 @@
-//
-// Copyright © 2025 Stream.io Inc. All rights reserved.
-//
-
-import Foundation
-import SwiftUI
-
-extension ChatChannelWatcherListController {
-    /// A wrapper object that exposes the controller variables in the form of `ObservableObject` to be used in SwiftUI.
-    public var observableObject: ObservableObject { .init(controller: self) }
-
-    /// A wrapper object for `ChatChannelWatcherListController` type which makes it possible to use the controller
-    /// comfortably in SwiftUI.
-    public class ObservableObject: SwiftUI.ObservableObject {
-        /// The underlying controller. You can still access it and call methods on it.
-        public let controller: ChatChannelWatcherListController
-
-        /// The channel members.
-        @Published public private(set) var watchers: [ChatUser] = []
-
-        /// The current state of the controller.
-        @Published public private(set) var state: DataController.State
-
-        /// Creates a new `ObservableObject` wrapper with the provided controller instance.
-        init(controller: ChatChannelWatcherListController) {
-            self.controller = controller
-            state = controller.state
-
-            controller.multicastDelegate.add(additionalDelegate: self)
-
-            watchers = controller.watchers
-        }
-    }
-}
-
-extension ChatChannelWatcherListController.ObservableObject: ChatChannelWatcherListControllerDelegate {
-    public func channelWatcherListController(
-        _ controller: ChatChannelWatcherListController,
-        didChangeWatchers changes: [ListChange<ChatUser>]
-    ) {
-        watchers = controller.watchers
-    }
-
-    public func controller(_ controller: DataController, didChangeState state: DataController.State) {
-        self.state = state
-    }
-}
